% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_win_vs_homestatus.R
\name{barplot_win_vs_homestatus}
\alias{barplot_win_vs_homestatus}
\title{Gives a quick barplot of number of wins.}
\usage{
barplot_win_vs_homestatus()
}
\value{
This function returns a split barplot with the number of games played,
split whether those wins were at a home event or away.
}
\description{
This will return a barplot of the number of wins vs traveling status.
}
\examples{
## Give a barplot of number of wins with split on traveling or home games
barplot_win_vs_homestatus()

}
