% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbbga24.R
\docType{data}
\name{cbbga24}
\alias{cbbga24}
\title{cbbga24}
\format{
A data frame with 6 variables: \code{Date}, \code{Away},
\code{A_score}, \code{Home}, \code{H_score}, \code{score_diff}
}
\source{
http://kenpom.com/cbbga24.txt
}
\usage{
cbbga24
}
\description{
Set of data containing Date of each game, away team name, away team score,
home team name, home team score, and difference between away team and
home team score.
}
\keyword{datasets}
