% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbbga_wins_23}
\alias{cbbga_wins_23}
\title{cbbga_wins_23}
\format{
A data frame with 7 variables: \code{team}, \code{win},
\code{loss}, \code{win_percentage}, \code{home_games}, \code{total_games},
\code{home_percentage}
}
\usage{
cbbga_wins_23
}
\description{
CSV containing team name, number of wins, number of losses,
win percentage, number of home games, total number of games played,
percentage of games played at home.
}
\keyword{datasets}
