% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_difference.R
\name{plot_score_difference}
\alias{plot_score_difference}
\title{This is my score difference histogram function}
\usage{
plot_score_difference(team)
}
\arguments{
\item{team}{this is the college team name}
}
\value{
This function returns a histogram of the score difference.
}
\description{
This is my score difference histogram function
}
\examples{
plot_score_difference("Southern Utah")

}
