% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_difference_box.R
\name{plot_score_difference_box}
\alias{plot_score_difference_box}
\title{This is my box plot function}
\usage{
plot_score_difference_box(team)
}
\arguments{
\item{team}{this is the college team name}
}
\value{
This function returns a boxplot of the score differences for the
specified team
}
\description{
This is my box plot function
}
\examples{
plot_score_difference_box("Southern Utah")

}
