% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_team.R
\name{Results_team}
\alias{Results_team}
\title{Gives specific results for a specified team}
\usage{
Results_team(team)
}
\arguments{
\item{team}{This is the desired team to get win and home game percentages}
}
\value{
This function returns a list containing the team results and home
game percentage.
}
\description{
This will calculate the win percentage and percentage of games the specified
team played at home with the data set cbbga24
}
\examples{
## Give win and home game percentages for 'Southern Utah'
Results_team('Southern Utah')

}
