% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTeamPlots.R
\name{runTeamPlots}
\alias{runTeamPlots}
\title{Team Plots App}
\usage{
runTeamPlots()
}
\description{
This function allows the TeamPlots shiny app to run. The app is a tool to give
basic team statistics for a provided team in the form of either a box plot or
a bar plot. If an incorrect name is given the program will character match all
the possible options so you can find the team you are looking for.
}
