% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_winpercentage_vs_homestatus.R
\name{scatterplot_winpercentage_vs_homestatus}
\alias{scatterplot_winpercentage_vs_homestatus}
\title{Gives a quick scatterplot of number of percentage of wins vs percentage of home games.}
\usage{
scatterplot_winpercentage_vs_homestatus()
}
\value{
This will return a scatterplot of the percentage of wins vs percentage
of home games.
}
\description{
This will return a scatterplot of the percentage of wins vs percentage of home games,
with colors referencing to the total number of games played by each team.
}
\examples{
## Give a scatterplot of win percentage vs home game percentage with colors referring
to the number of total games played.
barplot_win_vs_homestatus()

}
