% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_filter.R
\name{team_filter}
\alias{team_filter}
\title{Filter the data set ccbga24 by a specified team}
\usage{
team_filter(team)
}
\arguments{
\item{team}{This is the desired team to filter}
}
\value{
This function returns a tible with just games played by the specified team
}
\description{
This will filter the data set ccbga24, returning only games where the specified
team competes.
}
\examples{
## Filter for 'Southern Utah'
team_filter('Southern Utah')

}
