% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_function.R
\name{team_function}
\alias{team_function}
\title{this is the filter function}
\usage{
team_function(team)
}
\arguments{
\item{team}{this is the team name to filter by}
}
\value{
This returns a tibble for a data set for the specified team.
}
\description{
this is the filter function
}
\examples{
team_function('Southern Utah')


}
