% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_results.R
\name{team_results}
\alias{team_results}
\title{Gives results for a specified team}
\usage{
team_results(team)
}
\arguments{
\item{team}{This is the desired team to generate statistics for.}
}
\value{
This function returns a list containing the number of wins, losses,
win percentage, home games, total games played and percentage of games at home
for the specified team.
}
\description{
This will calculate the number of wins, losses, win percentage, number of games at home,
total games played and percentage of games played at home.
}
\examples{
## Give win and home game statistics for 'Southern Utah'
Results_team('Southern Utah')

}
