% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_win_perc.R
\name{Total_win_perc}
\alias{Total_win_perc}
\title{Gives quick team summary statistics}
\usage{
Total_win_perc()
}
\value{
This function returns a tibble containing each team, their win percentage,
home game percentage, and number of games played.
}
\description{
This will calculate the win percentage, percentage of games played at home, and
the number of games each team played for the data set cbbga24
}
\examples{
## Give quick team summary statistics
Total_win_perc

}
