% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{win_loss_ratio}
\alias{win_loss_ratio}
\title{win_loss_ratio Data Set}
\format{
A data frame with 4 variables: \code{team}, \code{win}, \code{loss} and \code{win_ratio}.
}
\usage{
win_loss_ratio
}
\description{
This data set contain iformation on the wins, losses, and win/loss ratios
of college basketball games played in 2023 for the 23-24 season
}
\keyword{datasets}
