% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbbga24}
\alias{cbbga24}
\title{Cbbga24}
\format{
A data frame with 6 variables: \code{Date}, \code{Team1}, \code{Score1},
\code{Team2}, \code{Score2}, and \code{Score_Difference}
}
\usage{
cbbga24
}
\description{
This data contains colleege basketball information in the year 2023 in the
2023-2024 season.
}
\keyword{datasets}
