% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_results.R
\name{team_results}
\alias{team_results}
\title{This is the team results function}
\usage{
team_results(team)
}
\arguments{
\item{team}{this is the college basketball team specified}
}
\value{
This function reports the number wins, losses, and win/loss ratio
as a tibble.
}
\description{
This is the team results function
}
\examples{
team_results('Southern Utah')


}
